Consolidate Array and Builder classes in array.h, builder.h. Remove arrow/types subdirectory
