+++ cpp/src/arrow/array-list-test.cc
@@ -27,10 +27,6 @@
 #include "arrow/builder.h"
 #include "arrow/test-util.h"
 #include "arrow/type.h"
-#include "arrow/types/construct.h"
-#include "arrow/types/list.h"
-#include "arrow/types/primitive.h"
-#include "arrow/types/test-common.h"
 #include "arrow/util/status.h"
 
 using std::shared_ptr;
