+++ cpp/src/arrow/array-primitive-test.cc
@@ -22,13 +22,11 @@
 
 #include "gtest/gtest.h"
 
+#include "arrow/array.h"
 #include "arrow/builder.h"
 #include "arrow/test-util.h"
 #include "arrow/type.h"
 #include "arrow/type_traits.h"
-#include "arrow/types/construct.h"
-#include "arrow/types/primitive.h"
-#include "arrow/types/test-common.h"
 #include "arrow/util/bit-util.h"
 #include "arrow/util/buffer.h"
 #include "arrow/util/status.h"
