+++ cpp/src/arrow/column-benchmark.cc
@@ -17,8 +17,8 @@
 
 #include "benchmark/benchmark.h"
 
+#include "arrow/array.h"
 #include "arrow/test-util.h"
-#include "arrow/types/primitive.h"
 #include "arrow/util/memory-pool.h"
 
 namespace arrow {
