+++ cpp/src/arrow/ipc/adapter.cc
@@ -33,10 +33,6 @@
 #include "arrow/table.h"
 #include "arrow/type.h"
 #include "arrow/types/construct.h"
-#include "arrow/types/list.h"
-#include "arrow/types/primitive.h"
-#include "arrow/types/string.h"
-#include "arrow/types/struct.h"
 #include "arrow/util/bit-util.h"
 #include "arrow/util/buffer.h"
 #include "arrow/util/logging.h"
