+++ cpp/src/arrow/ipc/ipc-adapter-test.cc
@@ -31,10 +31,6 @@
 #include "arrow/ipc/util.h"
 
 #include "arrow/test-util.h"
-#include "arrow/types/list.h"
-#include "arrow/types/primitive.h"
-#include "arrow/types/string.h"
-#include "arrow/types/struct.h"
 #include "arrow/util/bit-util.h"
 #include "arrow/util/buffer.h"
 #include "arrow/util/memory-pool.h"
